feat(error): added static error audio response
