+++ andy_api/api/api_routes.py
@@ -10,14 +10,35 @@ from flask import (
 )
 
 from . import speech_text_processing, dialogflow_andy
-from .intent_processing import intent_processing, error_fulfillment
-from .logging import create_intent_log, update_intent_log_with_audio_response, create_error_log, ERROR_TYPES
+from .intent_processing import intent_processing
+from .logging import (
+    create_intent_log,
+    update_intent_log_with_audio_response,
+    create_error_log,
+    ERROR_TYPES
+)
+from .api_route_helpers import get_response_error_return, get_static_error_audio
 
 bp = Blueprint('api', __name__, url_prefix='/api')
 
 
 @bp.route("/get-audio-response", methods=["POST"])
 def get_audio_response():
+    """Route for getting audio data from text.
+
+    Query Params:
+        session_id: the unique session ID to use with Andy.
+
+    Body:
+        Text that should be converted into audio.
+
+    Returns:
+        An HTTP response, with the data field containing the raw bytes of the
+        audio file. The data field will only be present if the status code of
+        the response is 200.
+
+    """
+    # pylint: disable=broad-except
     if request.method == "POST":
         session_id = request.args.get('session_id')
 
@@ -32,8 +53,8 @@ def get_audio_response():
         except Exception as e:
             err_msg = f"Error with text-to-speech: {e}"
             create_error_log(session_id, ERROR_TYPES.TTS, err_msg)
-            # TODO: replace with a return "error" response
-            raise
+            # Return a static audio response
+            return get_static_error_audio()
 
         # Log the audio response on a separate thread
         Thread(target=update_intent_log_with_audio_response(
@@ -78,6 +99,7 @@ def get_response():
         board_str (str): the state of the board, as a FEN string.
 
     """
+    # pylint: disable=broad-except
     if request.method == "POST":
         session_id = request.args.get('session_id')
         # grab board string from HTTP arguments
@@ -138,42 +160,3 @@ def get_response():
             'fulfillment_info': fulfillment_info,
             'board_str': updated_board_str,
         })
-
-
-def get_response_error_return(board_str):
-    """Returns a generic error response.
-
-    Args:
-        board_str: FEN representation of board from client.
-
-    Returns:
-        A JSON object that should be returned for the get-response route.
-
-        {
-            'response_text': str,
-            'board_str': str
-            'fulfillment_info': dict
-        }
-
-        response_text (str): the response generated by Andy, as text.
-        fulfillment_info (dict): the intent information detected from the user
-            and the game state.
-
-            {
-                'intent_name': str,
-                'success': boolean
-            }
-
-            intent_name (str): the name of the detected intent (always ERROR).
-            success (boolean): always False for an error.
-        board_str (str): the state of the board, as a FEN string.
-
-    """
-    # Get error fulfillment information
-    response_text, fulfillment_info = error_fulfillment.get_error_fulfillment()
-
-    return jsonify({
-        "response_text": response_text,
-        "fulfillment_info": fulfillment_info,
-        "board_str": board_str
-    })
